refactor: migrate ID fields from BIGINT to INTEGER and add Employee model

+++ models/permission.js
@@ -5,12 +5,12 @@ const Permission = sequelize.define(
     "Permission",
     {
         id: {
-            type: DataTypes.BIGINT,
+            type: DataTypes.INTEGER,
             primaryKey: true,
             autoIncrement: true,
         },
         resource: { type: DataTypes.STRING(100) },
-        actionId: { type: DataTypes.BIGINT, field: 'action_id' },
+        actionId: { type: DataTypes.INTEGER, field: 'action_id' },
         name: { type: DataTypes.STRING(150), unique: true },
         description: { type: DataTypes.TEXT },
     },
