+++ models/rolePermission.js
@@ -4,8 +4,8 @@ const sequelize = require("../db/sequelize");
 const RolePermission = sequelize.define(
     "RolePermission",
     {
-        roleId: { type: DataTypes.BIGINT, field: 'role_id', primaryKey: true },
-        permissionId: { type: DataTypes.BIGINT, field: 'permission_id', primaryKey: true },
+        roleId: { type: DataTypes.INTEGER, field: 'role_id', primaryKey: true },
+        permissionId: { type: DataTypes.INTEGER, field: 'permission_id', primaryKey: true },
     },
     {
         tableName: "role_permissions",
